[CLIENT-5128] Add Value::Unknown for uninterpreted particle types
Records containing particle types this client does not interpret
(legacy Java/C#/Python/Ruby/PHP/Erlang serializations, retired
DIGEST/LDT, unknown future codes) previously panicked the decoder.
They now decode as Value::Unknown(code, bytes) — the raw wire code
and raw payload, uninterpreted — in both the record read path and
msgpack-nested values.

- ParticleType keeps only interpreted types; From<u8> replaced by
  fallible try_from_u8; ParticleType::name_of(u8) renders legacy
  type names for display/error messages.
- Value::particle_type() now returns the raw u8 wire code.
- Unknown is strictly read-only: rejected as a bin value, inside
  lists/maps/CDT arguments, in expression literals, in query
  filters, and as a record key.

+++ aerospike-core/src/key.rs
@@ -94,7 +94,7 @@ impl Key {
         let mut hash = Ripemd160::new();
         hash.update(self.set_name.as_bytes());
         if let Some(ref user_key) = self.user_key {
-            hash.update([user_key.particle_type() as u8]);
+            hash.update([user_key.particle_type()]);
             user_key.write_key_bytes(&mut hash)?;
         } else {
             unreachable!();
